+++ test/_config/step2/db1/sc1/stage/fn003_st1/fn003_addone.py
@@ -1,0 +1,2 @@
+def addone_py(i):
+    return i+1